Litt omstrukturering etter tilbakemelding
