+++ src/main/kotlin/no/nav/familie/ks/sak/integrasjon/oppgave/OppgaveService.kt
@@ -137,7 +137,7 @@ class OppgaveService(
 
                 oppgaveErAvsluttet -> {}
                 else -> {
-                    val nyFrist = LocalDate.parse(gammelOppgave.fristFerdigstillelse!!).plus(forlengelse)
+                    val nyFrist = LocalDate.parse(gammelOppgave.fristFerdigstillelse).plus(forlengelse)
                     val nyOppgave = gammelOppgave.copy(fristFerdigstillelse = nyFrist?.toString())
                     integrasjonClient.oppdaterOppgave(nyOppgave.id!!, nyOppgave)
                 }
